move form parser fn and fix case
